cloud/requirements.txt
pymongo>=4.5.0
Flask>=2.3.3
Flask-Cors>=4.0.0
grpcio_tools==1.59.0